refactor graph check
