--- cat.graphql
@@ -1,17 +1,0 @@
-type Query {
-  spaceCats: [SpaceCat]
-}
-
-type SpaceCat {
-  id: ID!
-  name: String!
-  age: Int
-  missions: [Mission]
-}
-
-type Mission {
-  id: ID!
-  name: String!
-  description: String!
-}
-#type Query { spaceCats: [SpaceCat] } type SpaceCat { id: ID! name: String! age: Int missions: [Mission]} type Mission { id: ID! name: String! description: String!}
