--- crates/rover-client/src/operations/graph/async_check/async_check_mutation.graphql
@@ -1,26 +1,0 @@
-mutation GraphAsyncCheckMutation(
-  $input: CheckSchemaAsyncInput!,
-  $name: String!,
-  $graph_id: ID!
-) {
-  graph(id: $graph_id) {
-    variant(name: $name) {
-      submitCheckSchemaAsync(input: $input) {
-        __typename
-        ... on CheckRequestSuccess {
-          targetURL
-          workflowID
-        }
-        ... on InvalidInputError {
-          message
-        }
-        ... on PermissionError {
-          message
-        }
-        ... on PlanError {
-          message
-        }
-      }
-    }
-  }
-}
