+++ crates/rover-client/src/operations/graph/check_workflow/check_workflow_query.graphql
@@ -1,13 +1,14 @@
-mutation GraphCheckWorkflowQuery(
+query GraphCheckWorkflowQuery(
   $graph_id: ID!
   $workflow_id: ID!
 ) {
   graph(id: $graph_id) {
-    checkWorkflow(id: $checkWorkflowId) {
+    checkWorkflow(id: $workflow_id) {
       status
       tasks {
         __typename
         ... on OperationsCheckTask {
+          targetURL
           status
           result {
             changes {
