+++ crates/rover-client/src/operations/graph/check_workflow/mod.rs
@@ -2,4 +2,4 @@ mod runner;
 mod types;
 
 pub use runner::run;
-pub use types::CheckSchemaAsyncInput;
+pub use types::CheckWorkflowInput;
