+++ src/command/graph/check.rs
@@ -1,13 +1,18 @@
+use rover_client::operations::graph::check_workflow::{self, CheckWorkflowInput};
 use serde::Serialize;
 use structopt::StructOpt;
 
-use rover_client::operations::graph::async_check::{self, CheckSchemaAsyncInput};
-use rover_client::operations::graph::check::{self, GraphCheckInput};
+use rover_client::operations::graph::check::{self, CheckSchemaAsyncInput};
+//use rover_client::operations::graph::check::{self, GraphCheckInput};
 use rover_client::shared::{CheckConfig, GitContext, GraphRef, ValidationPeriod};
 
 use crate::command::RoverOutput;
 use crate::options::{CheckConfigOpts, GraphRefOpt, ProfileOpt, SchemaOpt};
 use crate::utils::client::StudioClientConfig;
+use crate::utils::parsers::{
+    parse_file_descriptor, parse_query_count_threshold, parse_query_percentage_threshold,
+    FileDescriptorType,
+};
 use crate::Result;
 
 #[derive(Debug, Serialize, StructOpt)]
@@ -25,7 +30,7 @@ pub struct Check {
     #[structopt(flatten)]
     config: CheckConfigOpts,
 
-    /// If the check should be run asynchronously
+    /// If the check should be run asynchronously and exit without waiting for check results
     #[structopt(long = "background")]
     background: bool,
 }
@@ -45,7 +50,7 @@ impl Check {
             "Checking the proposed schema against metrics from {}",
             &self.graph.graph_ref
         );
-        let res_check_async = async_check::run(
+        let workflow_res = check::run(
             CheckSchemaAsyncInput {
                 graph_ref: self.graph.graph_ref.clone(),
                 proposed_schema,
@@ -59,24 +64,17 @@ impl Check {
             &client,
         )?;
         if self.background {
-            // check fetch status
-            
-            // let res = check::run(
-            //     GraphCheckInput {
-            //         graph_ref: self.graph.clone(),
-            //         proposed_schema,
-            //         git_context,
-            //         config: CheckConfig {
-            //             query_count_threshold: self.query_count_threshold,
-            //             query_count_threshold_percentage: self.query_percentage_threshold,
-            //             validation_period: self.validation_period.clone(),
-            //         },
-            //     },
-            //     &client,
-            // )?;
-            // Ok(RoverOutput::CheckResponse(res))
+            Ok(RoverOutput::AsyncCheckResponse(workflow_res))
         } else {
-            Ok(RoverOutput::AsyncCheckResponse(res))
+            let check_res = check_workflow::run(
+                CheckWorkflowInput {
+                    graph_ref: self.graph.clone(),
+                    workflow_id: workflow_res.workflow_id.clone(),
+                },
+                &client,
+            )?;
+
+            Ok(RoverOutput::CheckResponse(check_res))
         }
     }
 }
